Reworked hashbang on fossilsource for portability reasons.
While I'm at it, I'll start adding an actual visible version.

I was getting sick of copying fossilsource over to some BSD and the
script not working purely because bash isn't in the same place as it
is under Linux. I should have learned this a while ago. Let's hope env
is always in the same place, as that's also complicated matters in the
past.